Refactoring launcher preference form into component so we can show in story mode.
