+++ rlbot_gui/gui/js/main-vue.js
@@ -1,6 +1,7 @@
 import AppearanceEditor from './appearance-editor-vue.js'
 import MutatorField from './mutator-field-vue.js'
 import BotCard from './bot-card-vue.js'
+import LauncherPreferenceModal from './launcher-preference-vue.js'
 
 const HUMAN = {'name': 'Human', 'type': 'human', 'image': 'imgs/human.png'};
 const STARTING_BOT_POOL = [
@@ -23,11 +24,6 @@ export default {
 
 		<b-navbar-nav class="ml-auto">
 			<b-spinner v-if="showProgressSpinner" variant="success" label="Spinning" class="mr-2"></b-spinner>
-			<b-button
-				v-b-modal.launcher-modal variant="dark">
-				<img class="platform-icon" src="imgs/steam.png" /> /
-				<img class="platform-icon" src="imgs/epic.png" />
-			</b-button>
 			<span id="sandbox-button-wrapper">
 				<b-button
 					@click="$router.replace('/sandbox')" variant="dark" class="ml-2"
@@ -225,28 +221,6 @@ export default {
 				<div><b-form-checkbox v-model="matchSettings.enable_lockstep">Enable Lockstep</b-form-checkbox></div>
 				<mutator-field label="Existing Match Behaviour" :options="matchOptions.match_behaviours" v-model="matchSettings.match_behavior" class="mt-3"></mutator-field>
 			</b-modal>
-			
-			<b-modal title="Preferred Rocket League Launcher" id="launcher-modal" size="md" hide-footer centered>
-				<div>
-					<b-form-group>
-						<b-form-radio v-model="launcherSettings.preferred_launcher" name="launcher-radios" value="steam">Steam</b-form-radio>
-						<b-form-radio v-model="launcherSettings.preferred_launcher" name="launcher-radios" value="epic">Epic Games</b-form-radio>
-						<b-form-checkbox
-							class="ml-4"
-							v-model="launcherSettings.use_login_tricks" 
-							:disabled="launcherSettings.preferred_launcher !== 'epic'">
-						
-							Use Epic Login Tricks <b-icon class="warning-icon" icon="exclamation-triangle-fill" v-b-tooltip.hover 
-							title="If you choose this, we'll do some fancy things to make sure your Epic account logs in successfully and loads your car + camera settings. 
-							It might look slightly weird on the Epic login server but they probably won't care."></b-icon>
-						</b-form-checkbox>
-					</b-form-group>
-				</div>
-				<div>
-					
-				</div>
-				<b-button variant="primary" class="mt-3" @click="saveLauncherSettings()">Save</b-button>
-			</b-modal>
 
 			<b-modal id="mutators-modal" title="Mutators" size="lg" hide-footer centered>
 
@@ -406,6 +380,10 @@ export default {
 				v-bind:path="appearancePath"
 				v-bind:map="matchSettings.map"
 				id="appearance-editor-dialog" />
+				
+		<b-modal title="Preferred Rocket League Launcher" id="launcher-modal" size="md" hide-footer centered>
+			<launcher-preference-modal modal-id="launcher-modal" />
+		</b-modal>
 
 	</div>
 
@@ -416,6 +394,7 @@ export default {
 		'appearance-editor': AppearanceEditor,
 		'mutator-field': MutatorField,
 		'bot-card': BotCard,
+		'launcher-preference-modal': LauncherPreferenceModal,
 	},
 	data () {
 		return {
@@ -456,7 +435,6 @@ export default {
 				auto_save_replay: false,
 				scripts: [],
 			},
-			launcherSettings: { preferred_launcher: 'epic', use_login_tricks: false },
 			randomMapPool: [],
 			packageString: null,
 			showSnackbar: false,
@@ -487,7 +465,6 @@ export default {
 			this.matchSettings.scripts = this.scriptPool.filter((val) => { return val.enabled });
 			eel.save_match_settings(this.matchSettings);
 			eel.save_team_settings(this.blueTeam, this.orangeTeam);
-			eel.save_launcher_settings(this.launcherSettings);
 
 			const blueBots = this.blueTeam.map((bot) => { return  {'name': bot.name, 'team': 0, 'type': bot.type, 'skill': bot.skill, 'path': bot.path} });
 			const orangeBots = this.orangeTeam.map((bot) => { return  {'name': bot.name, 'team': 1, 'type': bot.type, 'skill': bot.skill, 'path': bot.path} });
@@ -499,7 +476,7 @@ export default {
 
 			// start match asynchronously, so it doesn't block things like updating the background image
 			setTimeout(() => {
-				eel.start_match(blueBots.concat(orangeBots), this.matchSettings, this.launcherSettings);
+				eel.start_match(blueBots.concat(orangeBots), this.matchSettings);
 			}, 0);
 		},
 		killBots: function(event) {
@@ -671,11 +648,6 @@ export default {
 				this.resetMatchSettingsToDefault();
 			}
 		},
-		launcherSettingsReceived: function (launcherSettings) {
-			if (launcherSettings) {
-				Object.assign(this.launcherSettings, launcherSettings);
-			}
-		},
 		teamSettingsReceived: function (teamSettings) {
 			if (teamSettings) {
 				this.blueTeam = teamSettings.blue_team;
@@ -726,16 +698,11 @@ export default {
 			this.orangeTeam = bots.slice();
 			this.$bvModal.hide('recommendations-modal');
 		},
-		saveLauncherSettings: function() {
-			eel.save_launcher_settings(this.launcherSettings);
-			this.$bvModal.hide('launcher-modal');
-		},
 	},
 	created: function () {
 		eel.get_folder_settings()(this.folderSettingsReceived);
 		eel.get_match_options()(this.matchOptionsReceived);
 		eel.get_match_settings()(this.matchSettingsReceived);
-		eel.get_launcher_settings()(this.launcherSettingsReceived);
 		eel.get_team_settings()(this.teamSettingsReceived);
 
 		eel.get_language_support()((support) => {
@@ -753,12 +720,5 @@ export default {
 			self.downloadStatus = status;
 			self.downloadProgressPercent = progress;
 		}
-	},
-	watch: {
-		'launcherSettings.preferred_launcher': function(newVal) {
-			if (newVal === 'steam') {
-				this.launcherSettings.use_login_tricks = false;
-			}
-		}
-	},
+	}
 };
